+++ nonsoolmateClient/src/api/auth.ts
@@ -1,0 +1,13 @@
+const ACCESS_TOKEN = "token";
+
+export const getToken = () => {
+  return localStorage.getItem(ACCESS_TOKEN);
+};
+
+export const setToken = (token: string) => {
+  localStorage.setItem(ACCESS_TOKEN, token);
+};
+
+export const removeToken = () => {
+  localStorage.removeItem(ACCESS_TOKEN);
+};
