+++ nonsoolmateClient/src/api/axios.ts
@@ -1,18 +1,5 @@
 import axios, { AxiosInstance } from "axios";
-
-const ACCESS_TOKEN = "token";
-
-export const getToken = () => {
-  return localStorage.getItem(ACCESS_TOKEN);
-};
-
-export const setToken = (token: string) => {
-  localStorage.setItem(ACCESS_TOKEN, token);
-};
-
-export const removeToken = () => {
-  localStorage.removeItem(ACCESS_TOKEN);
-};
+import { getToken } from "./auth";
 
 export const client: AxiosInstance = axios.create({
   baseURL: `${import.meta.env.VITE_BASE_URL}`,
